refactor(core): modularize identity candidate datum resolver (#533)
Split locate (frame row / outliers), series identity, and shared context
types out of datum.ts. Keep a thin factory with attribute/lineage assembly
and re-export public helpers for existing test import paths.
