+++ packages/core/src/pipeline/candidate-construction/datum.ts
@@ -1,18 +1,27 @@
+/**
+ * Identity candidate datum resolver — thin factory + attribute assembly.
+ *
+ * Locate: datum-locate.ts · Series: datum-series.ts · Values: datum-values.ts
+ * Shared types: datum-types.ts · Represented-row filters: represented-rows.ts
+ */
 import type { CandidateBuildFacts, CandidateDatum } from "../../candidate-store.js";
 import type { LineageStore } from "../../identity.js";
-import type { GeometryBatch, Scene } from "../../scene.js";
+import type { Scene } from "../../scene.js";
 import type { CellValue, ColumnTable } from "../../table.js";
 import type { FacetPanelDef } from "../facets.js";
-import { candidateAutoMode } from "../frame-candidates-auto-mode.js";
 import type { LayerBinding, LayerFrame, MappedField, ResolvedColorScale } from "../types.js";
-import { ordinalSeriesRank, resolveCandidateLogicalValues } from "./datum-values.js";
-import { resolveCandidateFrameRow } from "./frame-row.js";
+import { candidateAutoMode } from "../frame-candidates-auto-mode.js";
+import { resolveCandidateLogicalValues } from "./datum-values.js";
+import { locateIdentityCandidate } from "./datum-locate.js";
+import { resolveIdentitySeriesAndMode } from "./datum-series.js";
+import type { IdentityCandidateResolveContext, LocatedIdentityCandidate } from "./datum-types.js";
 import type { CandidateIdentityIndex } from "./identity-index.js";
 import { filterRepresentedSourceRows } from "./represented-rows.js";
 
-// ---------------------------------------------------------------------------
-// Annotation context
-// ---------------------------------------------------------------------------
+// Public re-exports for characterization tests and external callers.
+export { resolveOutlierContext } from "./datum-locate.js";
+export { resolveCandidateSeries } from "./datum-series.js";
+
 function resolveAnnotationIntercepts(input: {
   frame: LayerFrame | undefined;
   primitiveIndex: number;
@@ -32,42 +41,6 @@ function resolveAnnotationIntercepts(input: {
   };
 }
 
-// ---------------------------------------------------------------------------
-// Candidate datum assembly
-// ---------------------------------------------------------------------------
-function assembleIdentityCandidateDatum(
-  ctx: IdentityCandidateResolveContext,
-  facts: CandidateBuildFacts,
-  located: LocatedIdentityCandidate,
-): CandidateDatum {
-  const attrs = resolveIdentityCandidateAttrs(ctx, facts, located);
-  const { xValue, yValue } = resolveCandidateLogicalValues({
-    annotationRule: attrs.annotationRule,
-    annotationX: attrs.annotationX,
-    annotationY: attrs.annotationY,
-    outlierSourceRow: located.outlierSourceRow,
-    sourceRow: located.sourceRow,
-    frame: located.frame,
-    frameRow: located.frameRow,
-    primitiveIndex: facts.primitiveIndex,
-    sourceValue: located.sourceValue,
-    xField: located.xField,
-    yField: located.yField,
-  });
-  return {
-    xValue,
-    yValue,
-    seriesId: attrs.group,
-    seriesRank: attrs.seriesRank,
-    sourceOrder: attrs.sourceOrder,
-    lineage: attrs.lineageKey,
-    autoMode: attrs.autoMode,
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Candidate attribute facts
-// ---------------------------------------------------------------------------
 interface IdentityCandidateAttrs {
   group: number;
   seriesRank: number;
@@ -79,212 +52,6 @@ interface IdentityCandidateAttrs {
   annotationY: CellValue;
 }
 
-// ---------------------------------------------------------------------------
-// Candidate attributes
-// ---------------------------------------------------------------------------
-function resolveIdentityCandidateAttrs(
-  ctx: IdentityCandidateResolveContext,
-  facts: CandidateBuildFacts,
-  located: LocatedIdentityCandidate,
-): IdentityCandidateAttrs {
-  const { group, seriesRank, autoMode } = resolveIdentitySeriesAndMode(ctx, facts, located);
-  const lineageAnn = resolveIdentityLineageAndAnnotation(ctx, facts, located, group);
-  return {
-    group,
-    seriesRank,
-    autoMode,
-    ...lineageAnn,
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Identity resolver context
-// ---------------------------------------------------------------------------
-interface IdentityCandidateResolveContext {
-  scene: Scene;
-  bindings: readonly LayerBinding[];
-  panelFrames: readonly (readonly LayerFrame[])[];
-  facetPanels: readonly FacetPanelDef[];
-  table: ColumnTable;
-  layerFields: readonly MappedField[][];
-  color: ResolvedColorScale | null;
-  fill: ResolvedColorScale | null;
-  lineage: LineageStore<number>;
-  getIdentityIndex: () => CandidateIdentityIndex;
-}
-
-// ---------------------------------------------------------------------------
-// Mapped field lookup
-// ---------------------------------------------------------------------------
-function resolveCandidateFieldChannels(fields: readonly MappedField[]): {
-  xField: string | undefined;
-  yField: string | undefined;
-  colorField: string | undefined;
-  fillField: string | undefined;
-} {
-  return {
-    xField: fields.find((field) => field.channel === "x")?.field,
-    yField: fields.find((field) => field.channel === "y")?.field,
-    colorField: fields.find((field) => field.channel === "color")?.field,
-    fillField: fields.find((field) => field.channel === "fill")?.field,
-  };
-}
-
-function makeSourceValueLookup(
-  table: ColumnTable,
-  sourceRow: number | null,
-): (field: string | undefined) => CellValue {
-  return (field) =>
-    sourceRow === null || field === undefined ? null : table.column(field)[sourceRow]!;
-}
-
-// ---------------------------------------------------------------------------
-// Lineage and annotation
-// ---------------------------------------------------------------------------
-function resolveIdentityLineageAndAnnotation(
-  ctx: IdentityCandidateResolveContext,
-  facts: CandidateBuildFacts,
-  located: LocatedIdentityCandidate,
-  group: number,
-): {
-  sourceOrder: number;
-  lineageKey: number;
-  annotationRule: boolean;
-  annotationX: CellValue;
-  annotationY: CellValue;
-} {
-  const {
-    sourceRow,
-    frame,
-    outlierSourceRow,
-    frameRow,
-    sourceRowsByGroup,
-    sourceRowsByGroupX,
-    sourceRowsByGroupBin,
-    sourceRowsByGroupY,
-  } = located;
-  const { annotationRule, annotationX, annotationY } = resolveAnnotationIntercepts({
-    frame,
-    primitiveIndex: facts.primitiveIndex,
-  });
-  const { sourceOrder, lineageKey } = resolveRepresentedSourceRows({
-    outlierSourceRow,
-    sourceRow,
-    group,
-    panelIndex: facts.panelIndex,
-    layerIndex: facts.layerIndex,
-    sourceRowsByGroup,
-    sourceRowsByGroupX,
-    sourceRowsByGroupBin,
-    sourceRowsByGroupY,
-    frame,
-    table: ctx.table,
-    frameRow,
-    lineage: ctx.lineage,
-    primitiveIndex: facts.primitiveIndex,
-  });
-  return { sourceOrder, lineageKey, annotationRule, annotationX, annotationY };
-}
-
-// ---------------------------------------------------------------------------
-// Located Candidate facts
-// ---------------------------------------------------------------------------
-interface LocatedIdentityCandidate {
-  sourceRow: number | null;
-  frame: LayerFrame | undefined;
-  outlierSourceRow: number | null;
-  frameRow: number;
-  derivedGroup: number;
-  sourceValue: (field: string | undefined) => CellValue;
-  xField: string | undefined;
-  yField: string | undefined;
-  colorField: string | undefined;
-  fillField: string | undefined;
-  seriesByRow: Map<string, number>;
-  sourceRowsByGroup: Map<string, number[]>;
-  sourceRowsByGroupX: Map<string, number[]>;
-  sourceRowsByGroupBin: Map<string, number[]>;
-  sourceRowsByGroupY: Map<string, number[]>;
-}
-
-// ---------------------------------------------------------------------------
-// Candidate location
-// ---------------------------------------------------------------------------
-function locateIdentityCandidate(
-  ctx: IdentityCandidateResolveContext,
-  facts: CandidateBuildFacts,
-): LocatedIdentityCandidate {
-  const {
-    seriesByRow,
-    sourceRowsByGroup,
-    sourceRowsByGroupX,
-    sourceRowsByGroupBin,
-    sourceRowsByGroupY,
-    frameGroups,
-  } = ctx.getIdentityIndex();
-  const fields = ctx.layerFields[facts.layerIndex] ?? [];
-  const sourceRow = facts.rowIndex;
-  const frame = ctx.panelFrames[facts.panelIndex]?.[facts.layerIndex];
-  const batch = ctx.scene.batches[facts.batchIndex]!;
-  const { outlierLocalRow, outlierSourceRow } = resolveOutlierContext({
-    frame,
-    batch,
-    primitiveIndex: facts.primitiveIndex,
-    facetPanel: ctx.facetPanels[facts.panelIndex],
-  });
-  const orderedGroups = frameGroups.get(`${facts.panelIndex}:${facts.layerIndex}`) ?? [0];
-  const { frameRow, derivedGroup } = resolveCandidateFrameRow({
-    frame,
-    batch,
-    primitiveIndex: facts.primitiveIndex,
-    orderedGroups,
-    outlierLocalRow,
-  });
-  const sourceValue = makeSourceValueLookup(ctx.table, sourceRow);
-  const { xField, yField, colorField, fillField } = resolveCandidateFieldChannels(fields);
-  return {
-    sourceRow,
-    frame,
-    outlierSourceRow,
-    frameRow,
-    derivedGroup,
-    sourceValue,
-    xField,
-    yField,
-    colorField,
-    fillField,
-    seriesByRow,
-    sourceRowsByGroup,
-    sourceRowsByGroupX,
-    sourceRowsByGroupBin,
-    sourceRowsByGroupY,
-  };
-}
-
-// ---------------------------------------------------------------------------
-// Boxplot outlier context
-// ---------------------------------------------------------------------------
-export function resolveOutlierContext(input: {
-  frame: LayerFrame | undefined;
-  batch: GeometryBatch;
-  primitiveIndex: number;
-  facetPanel: FacetPanelDef | undefined;
-}): { outlierLocalRow: number | null; outlierSourceRow: number | null } {
-  const { frame, batch, primitiveIndex, facetPanel } = input;
-  const outlierLocalRow =
-    frame?.box !== null && frame?.binding.layer.geom === "boxplot" && batch.kind === "points"
-      ? (frame?.box.outlierRow[primitiveIndex] ?? null)
-      : null;
-  const outlierSourceRow =
-    outlierLocalRow === null
-      ? null
-      : (facetPanel?.sourceRows?.[outlierLocalRow] ?? outlierLocalRow);
-  return { outlierLocalRow, outlierSourceRow };
-}
-
-// ---------------------------------------------------------------------------
-// Represented source rows
-// ---------------------------------------------------------------------------
 export function resolveRepresentedSourceRows(input: {
   outlierSourceRow: number | null;
   sourceRow: number | null;
@@ -346,9 +113,96 @@ export function resolveRepresentedSourceRows(input: {
   };
 }
 
-// ---------------------------------------------------------------------------
-// Identity datum resolution
-// ---------------------------------------------------------------------------
+function resolveIdentityLineageAndAnnotation(
+  ctx: IdentityCandidateResolveContext,
+  facts: CandidateBuildFacts,
+  located: LocatedIdentityCandidate,
+  group: number,
+): {
+  sourceOrder: number;
+  lineageKey: number;
+  annotationRule: boolean;
+  annotationX: CellValue;
+  annotationY: CellValue;
+} {
+  const {
+    sourceRow,
+    frame,
+    outlierSourceRow,
+    frameRow,
+    sourceRowsByGroup,
+    sourceRowsByGroupX,
+    sourceRowsByGroupBin,
+    sourceRowsByGroupY,
+  } = located;
+  const { annotationRule, annotationX, annotationY } = resolveAnnotationIntercepts({
+    frame,
+    primitiveIndex: facts.primitiveIndex,
+  });
+  const { sourceOrder, lineageKey } = resolveRepresentedSourceRows({
+    outlierSourceRow,
+    sourceRow,
+    group,
+    panelIndex: facts.panelIndex,
+    layerIndex: facts.layerIndex,
+    sourceRowsByGroup,
+    sourceRowsByGroupX,
+    sourceRowsByGroupBin,
+    sourceRowsByGroupY,
+    frame,
+    table: ctx.table,
+    frameRow,
+    lineage: ctx.lineage,
+    primitiveIndex: facts.primitiveIndex,
+  });
+  return { sourceOrder, lineageKey, annotationRule, annotationX, annotationY };
+}
+
+function resolveIdentityCandidateAttrs(
+  ctx: IdentityCandidateResolveContext,
+  facts: CandidateBuildFacts,
+  located: LocatedIdentityCandidate,
+): IdentityCandidateAttrs {
+  const { group, seriesRank, autoMode } = resolveIdentitySeriesAndMode(ctx, facts, located);
+  const lineageAnn = resolveIdentityLineageAndAnnotation(ctx, facts, located, group);
+  return {
+    group,
+    seriesRank,
+    autoMode,
+    ...lineageAnn,
+  };
+}
+
+function assembleIdentityCandidateDatum(
+  ctx: IdentityCandidateResolveContext,
+  facts: CandidateBuildFacts,
+  located: LocatedIdentityCandidate,
+): CandidateDatum {
+  const attrs = resolveIdentityCandidateAttrs(ctx, facts, located);
+  const { xValue, yValue } = resolveCandidateLogicalValues({
+    annotationRule: attrs.annotationRule,
+    annotationX: attrs.annotationX,
+    annotationY: attrs.annotationY,
+    outlierSourceRow: located.outlierSourceRow,
+    sourceRow: located.sourceRow,
+    frame: located.frame,
+    frameRow: located.frameRow,
+    primitiveIndex: facts.primitiveIndex,
+    sourceValue: located.sourceValue,
+    xField: located.xField,
+    yField: located.yField,
+  });
+  return {
+    xValue,
+    yValue,
+    seriesId: attrs.group,
+    seriesRank: attrs.seriesRank,
+    sourceOrder: attrs.sourceOrder,
+    lineage: attrs.lineageKey,
+    autoMode: attrs.autoMode,
+  };
+}
+
 function resolveIdentityCandidateDatum(
   ctx: IdentityCandidateResolveContext,
   facts: CandidateBuildFacts,
@@ -357,82 +211,6 @@ function resolveIdentityCandidateDatum(
   return assembleIdentityCandidateDatum(ctx, facts, located);
 }
 
-// ---------------------------------------------------------------------------
-// Series identity and mode
-// ---------------------------------------------------------------------------
-function resolveIdentitySeriesAndMode(
-  ctx: IdentityCandidateResolveContext,
-  facts: CandidateBuildFacts,
-  located: LocatedIdentityCandidate,
-): { group: number; seriesRank: number; autoMode: ReturnType<typeof candidateAutoMode> } {
-  const { sourceRow, frame, derivedGroup, sourceValue, colorField, fillField, seriesByRow } =
-    located;
-
-  const { group, seriesRank } = resolveCandidateSeries({
-    sourceRow,
-    derivedGroup,
-    seriesByRow,
-    panelIndex: facts.panelIndex,
-    layerIndex: facts.layerIndex,
-    color: ctx.color,
-    fill: ctx.fill,
-    colorField,
-    fillField,
-    sourceValue,
-  });
-  const autoMode = candidateAutoMode(
-    frame?.binding ?? ctx.bindings[facts.layerIndex]!,
-    facts.primitiveIndex,
-  );
-  return { group, seriesRank, autoMode };
-}
-
-// ---------------------------------------------------------------------------
-// Candidate series
-// ---------------------------------------------------------------------------
-export function resolveCandidateSeries(input: {
-  sourceRow: number | null;
-  derivedGroup: number;
-  seriesByRow: Map<string, number>;
-  panelIndex: number;
-  layerIndex: number;
-  color: ResolvedColorScale | null;
-  fill: ResolvedColorScale | null;
-  colorField: string | undefined;
-  fillField: string | undefined;
-  sourceValue: (field: string | undefined) => CellValue;
-}): { group: number; seriesRank: number } {
-  const {
-    sourceRow,
-    derivedGroup,
-    seriesByRow,
-    panelIndex,
-    layerIndex,
-    color,
-    fill,
-    colorField,
-    fillField,
-    sourceValue,
-  } = input;
-  const group =
-    sourceRow === null
-      ? derivedGroup
-      : (seriesByRow.get(`${panelIndex}:${layerIndex}:${sourceRow}`) ?? 0);
-  const seriesRank = ordinalSeriesRank({
-    color,
-    fill,
-    colorField,
-    fillField,
-    sourceRow,
-    sourceValue,
-    group,
-  });
-  return { group, seriesRank };
-}
-
-// ---------------------------------------------------------------------------
-// Identity datum resolver
-// ---------------------------------------------------------------------------
 export function createIdentityCandidateDatumResolver(input: {
   scene: Scene;
   bindings: readonly LayerBinding[];
